# server/docs/swagger.yaml
definitions:
...
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
...
  system.LoginLog:
    properties:
      ID:
        description: 主键ID
        type: integer
      createdAt:
        description: 创建时间
        type: string
      login_at:
        description: 登录时间
        type: string
      updatedAt:
        description: 更新时间
        type: string
      user_id:
        description: 用户ID
        type: integer
      username:
        description: 用户名
        type: string
    type: object
...
info:
  contact: {}
  description: 使用gin+vue进行极速开发的全栈开发基础平台
  title: Gin-Vue-Admin Swagger API接口文档
  version: v2.8.1
paths:
...
  /user/getUserLoginHistoryById:
    get:
      consumes:
      - application/json
      parameters:
      - description: 用户ID
        in: query
        name: userId
        type: integer
      - description: 开始时间（格式：2006-01-02）
        in: query
        name: start
        type: string
      - description: 结束时间（格式：2006-01-02）
        in: query
        name: end
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 查询成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/system.LoginLog'
                  type: array
                msg:
                  type: string
              type: object
        "400":
          description: 请求失败
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: 条件查询用户登录历史
      tags:
      - LoginLog
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: x-token
    type: apiKey
swagger: "2.0"
...